Update Github util
