--- app/src/back/utils/github_util.js
@@ -1,18 +1,0 @@
-const { exec } = require("child_process");
-
-function git_pull() {
-  exec("git pull", (error, stdout, stderr) => {
-    console.log("exec git pull");
-    if (error) {
-      console.log(`error: ${error.message}`);
-      return;
-    }
-    if (stderr) {
-      console.log(`stderr: ${stderr}`);
-      return;
-    }
-    console.log(`stdout: ${stdout}`);
-  });
-}
-
-module.exports = { git_pull };
